move context into App
